+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - January 10th, 2022
+🚨 Breaking 🚨
+### Added
+- Experimental Kotlin Multiplatform Library Plugin
+### Changed
+- Huge refactor, modules are now decoupled and much less rigid
+
 ## [0.3.3] - January 8th, 2022
 ### Changed
 - Fixed bug where documentation home page was being generated on plugin load